+++ resources/plugins/QuantME-CamundaPlugin/deployment/services/DeploymentPlugin.js
@@ -19,8 +19,8 @@ import ServiceDeploymentBindingModal from './ServiceDeploymentBindingModal';
 import { getRootProcess } from 'client/src/app/quantme/utilities/Utilities';
 
 import { createServiceInstance, uploadCSARToContainer } from 'client/src/app/quantme/deployment/OpenTOSCAUtils';
-import { bindUsingPull, bindUsingPush, getBindingType } from './BindingUtils';
-import { getServiceTasksToDeploy } from './DeploymentUtils';
+import { bindUsingPull, bindUsingPush, getBindingType } from 'client/src/app/quantme/deployment/BindingUtils';
+import { getServiceTasksToDeploy } from 'client/src/app/quantme/deployment/DeploymentUtils';
 
 const defaultState = {
   windowOpenDeploymentOverview: false,
